darwin: more external assembler
more workarounds: for gcc-mp-* we must use -mno-avx
and skip -march=native, as gcc has no integrated as to use.
we should point it to /usr/bin/as but this version could also be outdated.

with clang it is better: use -Wa,q instead, the integrated assembler
with full optimizations.